Process saved paragraphs if broken when disabled
If you disable lwc for a short section, that usually means that you just
don't want certain paragraphs expanded. This commit makes it so that
disabling lwc turns off the paragraph-related callbacks, but keeps the
page breaking-related callbacks enabled, so that widows/orphans will
still be expanded away if the page breaks while typesetting a small
disabled section.