Completely rewrite the `pre_output_filter` callback.
Previous versions of lwc would use the `pre_output_filter` callback to
check to see if the `\outputpenalty` indicated that a widow or orphan
line was present on the current page. If so, it would replace one
paragraph with its expanded version, then move the last line of the page
onto the next page.

Complicating matters were `\insert`s---for these, lwc had to store the
inserts of the last line so that it could manually look into some
internal node lists to shuffle the appropriate nodes to the next page.
This was very tricky to do properly, but it generally worked okay.

Unfortunately, LuaTeX offers no similar interfaces to adjust any
`\mark`s, so lwc had no way to fix these if they were moved to the next
page. Additionally, moving the last line and its inserts was very
fragile, and a frequent source of bugs.

With this commit, lwc now completely avoids moving the last line, and
lets TeX handle it as usual. The key to making this work is to set a
“no-op” output routine sometimes, although the details are quite tricky.

Whenever LuaTeX adds content to the main vertical list, it first calls
the `buildpage_filter` callback, then it attempts to “build” the page.
To intercept this, we try to “build” the page _inside_ of the
`buildpage_filter` callback. This seems pointless, but it let's us set
`\holdinginserts` to a positive value, which forces TeX to not process
any pending `\insert`s By the time that `pre_output_filter` is called,
it's too late to do this, so we need to do it now.

Then, we set our `pre_output_filter` callback to only run if it was
called from inside of our `buildpage_filter` callback. At this point, we
unconditionally set a “no-op” output routine; that is, something
approximately like `\output={\unvbox255 \penalty\outputpenalty}`. We
need to do this because if the “real” output routine is called, it won't
see any inserts on the page, so they will be lost entirely.

Inside of the `pre_output_filter` callback, we check the
`\outputpenalty` to see if it indicates a widow or orphan line, and if
so, expand one paragraph on the page, just as we had done before.
However, unlike before, we don't need to do anything else---after the
“no-op” output routine is called, TeX will immediately build the page
again and call the “real” output routine, but this time the page will
be one line longer so the widow or orphan line will be automatically
avoided.

We will also need to test this some more before an actual release, but
all but 2 of the automated tests are currently passing, so despite the
large change, it seems to be working well.

Unfinished as of yet: documentation, ConTeXt support, `nobreak` option
(should be safe to remove entirely?), `tests/latex/footnotes-many`
failing (probably spurious), etc.