Pin Readme manual link to v3.0.0
Lua-widow-control was broken with ConTeXt when the latest version
(v3.0.1) was released, and because the manual uses ConTeXt, this means I
had to disable lua-widow-control when I compiled the manual. The manual
was correct for v3.0.0, so I copied that into the CTAN/TeX Live release
for v3.0.1, but the README.md points to the latest release with the
broken manual.

Until I make a new release with a fixed manual, let's just pin the link
in README.md to the v3.0.0 manual.

Fixes #46.